requirements.txt
openai
python-dotenv
colorama
flask